taxease_server/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
distro==1.9.0
Django==6.0.3
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
openai==2.29.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
